covid_sort_django/requirements.txt
Django==4.1.2
matplotlib==3.5.2
numpy==1.22.3
Pillow==9.0.1
torch==1.12.0
torchvision==0.13.0
